+++ fossbot_lib/dummy_robot/fossbot.py
@@ -15,7 +15,7 @@ class FossBot(robot_interface.FossBotInterface):
         """
         print('Just moving.')
 
-    def move_distance(self, dist: int, direction: str = "forward") -> None:
+    def move_distance(self, dist: float, direction: str = "forward") -> None:
         '''
         Moves to input direction (default == forward) a specified - input distance (cm).
         Param: dist: the distance to be moved (in cm).
@@ -41,7 +41,7 @@ class FossBot(robot_interface.FossBotInterface):
         print('Sleeping...')
 
     # moving forward
-    def move_forward_distance(self, dist: int) -> None:
+    def move_forward_distance(self, dist: float) -> None:
         '''
         Moves robot forward input distance.
         Param: dist: the distance (cm) to be moved by robot.
@@ -61,7 +61,7 @@ class FossBot(robot_interface.FossBotInterface):
         print('Moving forward.')
 
     # moving reverse
-    def move_reverse_distance(self, dist: int) -> None:
+    def move_reverse_distance(self, dist: float) -> None:
         '''
         Moves robot input distance in reverse.
         Param: dist: the distance (cm) to be moved by robot.
